rust: remove a weak symbol from jemalloc
je_malloc_conf is exported, weak, and initialized to 0. This generates a
relocation in BSS, which prelink doesn't fix up. This causes it to crash
due to accessing the wrong address.

The real fix would be in prelink, but there only seems to be one of
these problematic relocations in all the shared objects that get
prelinked. Prelink is quite difficult to understand and modify.

Also enable debug symbols in jemalloc.